requirements.txt
matplotlib
munch
numpy
pandas
pillow
pyyaml
scikit-learn
timm
pytorch
torchmetrics
torchvision
tqdm